Move ScanChainConfig into src/main/scala
This migrates the scan-chain-config utility into the main source
directory. Fat jar assembly is updated to work out of this directory.
