+++ project/plugins.sbt
@@ -1,1 +1,2 @@
 addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "1.0.0")
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.5")
