--- scan-chain-config/.gitignore
@@ -1,1 +1,0 @@
-utils
